# api/openapi.yaml
openapi: 3.0.3
info:
  title: messageFeed API
  version: 0.2.0
  description: Phase two and early phase four API contract for the feed reading workflow.
servers:
  - url: http://localhost:60001
    description: Local API server
paths:
...
  /api/v1/items:
    get:
      summary: List feed items
      description: Returns timeline items with source metadata and user item state.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: source_id
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: is_read
          in: query
          schema:
            type: boolean
        - name: is_favorite
          in: query
          schema:
            type: boolean
        - name: is_hidden
          in: query
          schema:
            type: boolean
        - name: include_hidden
          in: query
          description: Hidden items are excluded by default unless this is true or is_hidden is set explicitly.
          schema:
            type: boolean
            default: false
        - name: order
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
      responses:
        '200':
          $ref: '#/components/responses/ItemListResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/items/{id}:
    get:
      summary: Get feed item detail
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Feed item detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/items/{id}/read:
    post:
      summary: Set item read state
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                is_read:
                  type: boolean
                  default: true
      responses:
        '200':
          $ref: '#/components/responses/UserItemStateResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/items/{id}/favorite:
    post:
      summary: Set item favorite state
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                is_favorite:
                  type: boolean
                  default: true
      responses:
        '200':
          $ref: '#/components/responses/UserItemStateResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/items/{id}/hide:
    post:
      summary: Set item hidden state
      parameters:
        - $ref: '#/components/parameters/ID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                is_hidden:
                  type: boolean
                  default: true
      responses:
        '200':
          $ref: '#/components/responses/UserItemStateResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/feed/timeline:
    get:
      summary: List subscribed-source timeline items
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: source_id
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: is_read
          in: query
          schema:
            type: boolean
        - name: is_favorite
          in: query
          schema:
            type: boolean
        - name: is_hidden
          in: query
          schema:
            type: boolean
        - name: include_hidden
          in: query
          schema:
            type: boolean
            default: false
        - name: order
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
      responses:
        '200':
          $ref: '#/components/responses/ItemListResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/feed/recommendations:
    get:
      summary: List recommendation feed prototype items
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: source_id
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: refresh
          in: query
          schema:
            type: boolean
            default: false
        - name: order
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
      responses:
        '200':
          $ref: '#/components/responses/ItemListResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/feed/view-mode:
    get:
      summary: Get current feed view mode
      responses:
        '200':
          description: Current feed view mode
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FeedViewPreference'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    put:
      summary: Save current feed view mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - view_mode
              properties:
                view_mode:
                  $ref: '#/components/schemas/FeedViewMode'
      responses:
        '200':
          description: Saved feed view mode
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FeedViewPreference'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIResponse'
    ServiceUnavailable:
      description: Service dependency is unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIResponse'
    ItemListResult:
      description: Feed items
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/APIResponse'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ItemList'
    UserItemStateResult:
      description: Item state
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/APIResponse'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/UserItemState'
...
  schemas:
    APIResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: 0 for success; HTTP-like code for current errors.
          example: 0
        message:
          type: string
          example: success
        request_id:
          type: string
        trace_id:
          type: string
          description: OpenTelemetry trace id for correlating API responses with logs and traces.
        data:
          nullable: true
...
    ItemList:
      type: object
      required:
        - items
        - total
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    Item:
      type: object
      required:
        - id
        - source_id
        - title
        - url
        - normalized_url
        - fetched_at
        - is_read
        - is_favorite
        - is_hidden
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        source_id:
          type: integer
          format: int64
        source_name:
          type: string
        title:
          type: string
        url:
          type: string
          format: uri
        normalized_url:
          type: string
          format: uri
        raw_guid:
          type: string
        content_hash:
          type: string
        summary:
          type: string
        content_snippet:
          type: string
          description: Raw feed snippet. Frontend must not render it with v-html unless sanitized.
        content_text:
          type: string
          description: Plain text derived from content_snippet for safe default display.
        content_html:
          type: string
          description: Current raw HTML payload for iframe/srcdoc rendering; sanitize before inline rendering.
        author:
          type: string
        published_at:
          type: string
          format: date-time
          nullable: true
        fetched_at:
          type: string
          format: date-time
        is_read:
          type: boolean
        read_at:
          type: string
          format: date-time
          nullable: true
        is_favorite:
          type: boolean
        favorited_at:
          type: string
          format: date-time
          nullable: true
        is_hidden:
          type: boolean
        hidden_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserItemState:
      type: object
      required:
        - id
        - user_id
        - item_id
        - is_read
        - is_favorite
        - is_hidden
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        item_id:
          type: integer
          format: int64
        is_read:
          type: boolean
        read_at:
          type: string
          format: date-time
          nullable: true
        is_favorite:
          type: boolean
        favorited_at:
          type: string
          format: date-time
          nullable: true
        is_hidden:
          type: boolean
        hidden_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FeedViewPreference:
      type: object
      required:
        - user_id
        - view_mode
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        view_mode:
          $ref: '#/components/schemas/FeedViewMode'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FeedViewMode:
      type: string
      enum:
        - timeline